requirements.txt
winstats
pyrebase
pyyaml